backend/requirements.txt
Django>=5.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary 